# Title: Frame Rate Independent Movement
# Body:
As we are trying to move our sprite on the screen we encounter a
problem: the speed of its movement depends on the cpu clock,
which differ based on the machine that runs our game.

In order to achieve independent movement we've made a few changes.

# Game.cpp:
Added clock and deltaTime members, and calculateDeltaTime method to
calculate the time that passed from the last call. This will be saved
in the deltaTime member to calculate the number of pixels to move in
the current frame.

# main.cpp:
After each iteration we call calculateDeltaTime to restart the time.